requirements.txt
google-generativeai>=0.4.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
scipy>=1.10.0
jupyter>=1.0.0
python-dotenv>=1.0.0 